packaging improvements
made some improvements to make it easier to package and distribute the app. Most notably switched from electron forge to electron build for the packaging process.

### IMPORTANT NOTE:

When packaging place the llm folder outside the app directory and somewhere like the desktop, adjust package.json to include the new llm directory